+++ Editor/Editors/UImGuiEditor.cs
@@ -1,6 +1,7 @@
 using ImGuiNET;
 using System.Text;
 using UImGui.Platform;
+using UImGui.Renderer;
 using UnityEditor;
 using UnityEngine;
 
@@ -21,6 +22,10 @@ namespace UImGui.Editor
 		private SerializedProperty _style;
 		private SerializedProperty _cursorShapes;
 		private readonly StringBuilder _messages = new StringBuilder();
+
+		private bool usingImNodes = true;
+		private bool usingImGuizmo = true;
+		private bool usingImPlot = true;
 
 		public override void OnInspectorGUI()
 		{
@@ -74,11 +79,29 @@ namespace UImGui.Editor
 			_shaders = serializedObject.FindProperty("_shaders");
 			_style = serializedObject.FindProperty("_style");
 			_cursorShapes = serializedObject.FindProperty("_cursorShapes");
+
+#if UIMGUI_REMOVE_IMNODES
+			usingImNodes = false;
+#endif
+#if UIMGUI_REMOVE_IMGUIZMO
+			usingImGuizmo = false;
+#endif
+#if UIMGUI_REMOVE_IMPLOT
+			usingImPlot = false;
+#endif
 		}
 
 		private void CheckRequirements()
 		{
-			EditorGUILayout.LabelField("ImGUI Version: " + ImGui.GetVersion());
+			var textImGui = $"ImGUI: {ImGui.GetVersion()}";
+			var textImNodes = $"ImNodes: { (usingImNodes ? "0.4 - 2021-07-09" : "disabled") }";
+			var textImGuizmo = $"ImGuizmo: { (usingImGuizmo ? "?? - 2021-07-09" : "disabled") }";
+			var textImPlot = $"ImPlot: { (usingImPlot ? "0.10 - 2021-07-09" : "disabled") }";
+
+			EditorGUILayout.LabelField(textImGui);
+			EditorGUILayout.LabelField(textImNodes);
+			EditorGUILayout.LabelField(textImGuizmo);
+			EditorGUILayout.LabelField(textImPlot);
 			EditorGUILayout.Space();
 
 			_messages.Clear();
@@ -91,6 +114,13 @@ namespace UImGui.Editor
 			{
 				_messages.AppendLine("Must assign a RenderFeature when using the URP.");
 			}
+
+#if !UNITY_2020_1_OR_NEWER
+			if ((RenderType)_renderer.enumValueIndex == RenderType.Mesh)
+			{
+				_messages.AppendLine("Use procedural.");
+			}
+#endif
 
 			SerializedProperty configFlags = _initialConfiguration.FindPropertyRelative("ImGuiConfig");
 			if (!PlatformUtility.IsAvailable((InputType)_platform.enumValueIndex))