+++ Source/Renderer/RendererProcedural.cs
@@ -1,5 +1,4 @@
 ﻿using ImGuiNET;
-using NumericsConverter;
 using System;
 using System.Runtime.InteropServices;
 using UImGui.Assets;
@@ -9,7 +8,6 @@ using Unity.Collections.LowLevel.Unsafe;
 using UnityEngine;
 using UnityEngine.Assertions;
 using UnityEngine.Rendering;
-using NVector4 = System.Numerics.Vector4;
 using Object = UnityEngine.Object;
 
 // TODO: switch from using ComputeBuffer to GraphicsBuffer
@@ -32,7 +30,6 @@ namespace UImGui.Renderer
 		private readonly TextureManager _textureManager;
 
 		private readonly MaterialPropertyBlock _materialProperties = new MaterialPropertyBlock();
-		private readonly int[] _drawArgs = new int[] { 0, 1, 0, 0, 0 }; // Used to build argument buffer.
 
 		private Material _material;
 
@@ -78,10 +75,10 @@ namespace UImGui.Renderer
 
 		public void RenderDrawLists(CommandBuffer cmd, ImDrawDataPtr drawData)
 		{
-			System.Numerics.Vector2 fbSize = drawData.DisplaySize * drawData.FramebufferScale;
+			Vector2 fbSize = drawData.DisplaySize * drawData.FramebufferScale;
 
 			// Avoid rendering when minimized.
-			if (fbSize.X <= 0f || fbSize.Y <= 0f || drawData.TotalVtxCount == 0) return;
+			if (fbSize.x <= 0f || fbSize.y <= 0f || drawData.TotalVtxCount == 0) return;
 
 			Constants.UpdateBuffersMarker.Begin();
 			UpdateBuffers(drawData);
@@ -90,7 +87,7 @@ namespace UImGui.Renderer
 			cmd.BeginSample(Constants.ExecuteDrawCommandsMarker);
 
 			Constants.CreateDrawComandsMarker.Begin();
-			CreateDrawCommands(cmd, drawData, fbSize.ToUnity());
+			CreateDrawCommands(cmd, drawData, fbSize);
 			Constants.CreateDrawComandsMarker.End();
 
 			cmd.EndSample(Constants.ExecuteDrawCommandsMarker);
@@ -173,14 +170,18 @@ namespace UImGui.Renderer
 				_indexBuffer.SetData(idxArray, 0, idxOf, idxArray.Length);
 
 				// Arguments for indexed draw.
-				_drawArgs[3] = vtxOf; // Base vertex location.
-				for (int i = 0, iMax = drawList.CmdBuffer.Size; i < iMax; ++i)
+				for (int meshIndex = 0, iMax = drawList.CmdBuffer.Size; meshIndex < iMax; ++meshIndex)
 				{
-					ImDrawCmdPtr cmd = drawList.CmdBuffer[i];
-					_drawArgs[0] = (int)cmd.ElemCount; // Index count per instance.
-					_drawArgs[2] = idxOf + (int)cmd.IdxOffset; // Start index location.
-					_argumentsBuffer.SetData(_drawArgs, 0, argOf, 5);
-
+					ImDrawCmdPtr cmd = drawList.CmdBuffer[meshIndex];
+					var drawArgs = new int[]
+					{
+						(int)cmd.ElemCount,
+						1,
+						idxOf + (int)cmd.IdxOffset,
+						vtxOf,
+						0
+					};
+					_argumentsBuffer.SetData(drawArgs, 0, argOf, 5);
 					argOf += 5; // 5 int for each command.
 				}
 				vtxOf += vtxArray.Length;
@@ -191,10 +192,10 @@ namespace UImGui.Renderer
 		private void CreateDrawCommands(CommandBuffer cmd, ImDrawDataPtr drawData, Vector2 fbSize)
 		{
 			IntPtr prevTextureId = IntPtr.Zero;
-			NVector4 clipOffst = new NVector4(drawData.DisplayPos.X, drawData.DisplayPos.Y,
-				drawData.DisplayPos.X, drawData.DisplayPos.Y);
-			Vector4 clipScale = new Vector4(drawData.FramebufferScale.X, drawData.FramebufferScale.Y,
-				drawData.FramebufferScale.X, drawData.FramebufferScale.Y);
+			Vector4 clipOffst = new Vector4(drawData.DisplayPos.x, drawData.DisplayPos.y,
+				drawData.DisplayPos.x, drawData.DisplayPos.y);
+			Vector4 clipScale = new Vector4(drawData.FramebufferScale.x, drawData.FramebufferScale.y,
+				drawData.FramebufferScale.x, drawData.FramebufferScale.y);
 
 			_material.SetBuffer(_verticesID, _vertexBuffer); // Bind vertex buffer.
 
@@ -219,7 +220,7 @@ namespace UImGui.Renderer
 					else
 					{
 						// Project scissor rectangle into framebuffer space and skip if fully outside.
-						Vector4 clipSize = (drawCmd.ClipRect - clipOffst).ToUnity();
+						Vector4 clipSize = drawCmd.ClipRect - clipOffst;
 						Vector4 clip = Vector4.Scale(clipSize, clipScale);
 
 						if (clip.x >= fbSize.x || clip.y >= fbSize.y || clip.z < 0f || clip.w < 0f) continue;